Fixed bugs, and added better scalability to the surface area
git-svn-id: svn://t2.unl.edu/brian/HadoopViz/trunk@4007 1fda6f16-4416-0410-bfd6-867a04880151

+++ HadoopDropper/HadoopDropper.h
@@ -42,7 +42,7 @@ private:
     GLfloat* m_floorVertices;
     int m_numVertices;
     int m_numTiles;
-    GLubyte* m_tileIndices;
+    GLuint* m_tileIndices;
     GLubyte* m_floorColors;
 };
